fix: prevent showing gizmos on entities added to camera or player
